+++ macserial/src/macserial.h
@@ -33,7 +33,7 @@ uint32_t arc4random_uniform(uint32_t upper_bound) __attribute__((weak));
 #define SERIAL_YEAR_OLD_MAX 2012
 
 #define SERIAL_YEAR_NEW_MIN 2010
-#define SERIAL_YEAR_NEW_MAX 2019
+#define SERIAL_YEAR_NEW_MAX 2020
 
 #define SERIAL_COPY_MIN 1
 #define SERIAL_COPY_MAX 34